Trim file names before adding them to qml components and scripts
Fix the issue with the trailing spaces in the file paths of
components and scripts in the json output. This specificly fixes
androiddeployqt behavior since it was not able to find qml
components mentioned in the qmlimportscanner's output and skip
qml modules, because paths contained trailing spaces.

Amends 2cc85e60d4313bd845bfdc233e5e75a8b7ccb01f

Pick-to: 5.15.11
Fixes: QTBUG-107255
Change-Id: I20cc1b724aaea3a76932e82a6dc65eccbef3576d
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>