Massage qt-add-qml-module.qdoc to be easier to read
Split the huge blob-section "Arguments" describing the arguments of the
qt_add_qml_module command into groups (as \section2 elements).

Sort the newly obtained groups around to have the probably most needed
ones on top and the more special cased ones on the bottom.

Add an example where DEPENDENCIES entries are needed, such that it is
easier findable and understandable when it is typically needed.

Fixes: QTBUG-107092
Change-Id: I52ff94d038645e1383c56f6ff384b55627e800b5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>