qmltc: Remove broken alias optimization
Remove the recursive alias resolution mechanism. It tried to resolve
aliases recursively but only knew the id's of the current component.
This means that an alias pointing to a property defined in another
component (e.g. a different qml file) was resolved using the
id-to-scope map of the component containing the alias property.
This leads to unresolved aliases at best and endless loops when the
aliased property was itself an alias pointing to some property also
existing in the original component (see
QmltcTests/ComponentWithAlias{1..3}.qml for an example).

Also added the reproducer in the qmltc tests.

Fixes: QTBUG-107533
Fixes: QTBUG-107534
Change-Id: Idbf0cff3e45213bc7fbb4a98a9393c6754b00046
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 259ff52e64cd8725bfaa7330a33f8d36456e8b1a)