Fix baseline execution with native style
Don't override the native style detection of Qt Quick Controls, let the
framework decide, and only print information about what to expect.
This fixes running the base line tests with the native windows style,
and has the benefit that we will see mismatches if the native style on
a platform changes.

Change-Id: I8e9e855e3acb6a38d46ae84a1ec2582d4eb22b0f
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>