QmlCompiler: early return from array bindings
List in qml should not include an object whose scope types is attached or grouped.
In case an attached or grouped type is a member of a list, they don't
have a separate scope, but only one scope is created per name. This
causes a crash as creating bindings for array members since it is not
known if any of the child scope is omitted or not. To prevent this,
simply early return if any of the child scopes include an object
definition that is attached object or group type.

Fixes: QTBUG-107250
Change-Id: I08fdfb43d766569c4dfefc50135f98f74d25bad2
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>