Remove dead code
Since the Windows nativestyle SpinBox never uses a NativeStyle.SpinBox
as a background, there is no need in having the bool __nativeBackground
property (the background is always an Item).

We can then remove/simplify all code that uses the __nativeBackground
property, since it was already always false before this change.

Also removes this warning:

  SpinBox.qml:15:5: Unable to assign [undefined] to int

Pick-to: 6.4 6.2
Change-Id: I2b9e44617884ba4523fbe17417aaccc432c1cb50
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>