iOS Style: Make QQuickIOSStyle a singleton and private API
We weren't really supporting an attached API for the iOS style,
even though QQuickIOSStyle was a QQuickAttachedObject.

The QQuickIOSStyle had two properties: theme and url. We were using
the read-only IOS.theme property to select the source for the assets
based on the theme (light/dark), but this is not needed anymore as
we are now using the new Qt.styleHints.appearance property instead.
The url property was also not configurable and the controls
implementations were using it to access the path of the image assets.

As both of these properties were not settable and propagatable, and
we don't intend to make them as such, it doesn't make sense to have an
attached API for this. So make the QQuickIOSStyle a QObject and a
singleton. Remove the theme property.

Since the API doesn't need to be public, move it to the private
iOS.impl module.

Change-Id: I197aadba2032a30d97a3d9a6698b74aa8709b6f9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>