Fix bug in FolderListModel's FileInfoThread
When the FileDialog in the goUp test is first created, it sets a bunch
of properties, two of which are sortCaseSensitive and nameFilters.
These cause FileInfoThread's sortUpdate and folderUpdate flags to be set
to true, respectively.

When FileInfoThread::getFileInfos() ran, it would see that folderUpdate
was set to true, run related code, and then unset folderUpdate. However,
sortUpdate would not be unset. This meant that the next time
getFileInfos() was called, it would effectively reset the model when it
didn't need to. This caused the currentIndex we set on
fileDialogListView to be ignored because the ListView set it to 0 as a
result of the insertion and removal of rows from the model caused by
QQuickFolderListModelPrivate::_q_sortFinished().

This patch fixes the issue by:

- Introducing an enum to manage the different types of updates
  requested, which also makes the code a little easier to understand.
- Unsetting the various flags whenever needUpdate is unset.

It also adds categorized logging to make debugging such issues in the
future easier.

Task-number: QTBUG-101488
Pick-to: 6.4
Change-Id: Iaf53509d3f028fd2e1385a2597898e8e0327aedb
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>