SelectionRectangle: ask TableView for acceptance before starting a selection
Add a new virtual function to QQuickSelectable that lets
SelectionRectangle ask TableView up front if it's OK
to start a new selection at the given pos. This will
be useful once TableView adds its own pointer/drag
handler for resizing columns, and will be used to
avoid a situation where a selection starts at the
same time as a resize.

Task-number: QTBUG-106792
Change-Id: Ib0ed2d46802a05b717844ea6278ceef03f28840e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>