TextEdit: deal with scrolling backwards in plain text
When the user scrolls backwards, we have a backwards iteration loop to
find the block that is at or above the top of the viewport. But if the
user tried to jump-scroll by some distance, boundingRect() and
nodeOffset were staying constant, so the iteration tended to go all the
way back to the first block. We need to update nodeOffset at each step.

Also make test failures in largeTextObservesViewport() somewhat easier
to troubleshoot.

Amends cd083920b3b4f3a1ed7f2297058cf0d110d7cf10

Pick-to: 6.4
Fixes: QTBUG-104526
Change-Id: I590f59baf7137f78c4c294aa2bfbb040c80d4874
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>