QQmlTreeModelToTableModel: move logic from modelLayoutChanged() to modelLayoutAboutToBeChanged()
The QQmlTreeModelToTableModel contains a QList of items that
converts the source model (which typically is a tree model)
into a flat list of items that can be be displayed in a TableView.

When the source model is changing the layout (e.g when doing
a sort), we handle this in QQmlTreeModelToTableModel by listening
to the modelLayoutChanged() signal. The problem is that by the
time we receive that signal, all the TreeItems in the list of
items will have their QPersistentModelIndex updated, before we
have updated the layout of the list to be in sync. Since functions
like itemIndex(index) depends on the list model and the tree model
to be in sync, it cannot be trusted at this point.

To handle this problem, we need to divide the current logic into
two functions. The first part will remove all the affected rows
from the list before the source model has changed (while the models
are still in sync), from modelLayoutAboutToBeChanged(). Then we
add them back again once the source model has finished updating the
layout, which is done from modelLayoutChanged().

This will fix a bug in TreeView, where the view will not reflect
the sorting changes done to a QSortFilterProxyModel applied on
a QAbstractItemModel.

Fixes: QTBUG-103877
Change-Id: I9acfc3b7b9a79c2d16e1bc63e56b1d6ea5a53897
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 96e70f23666123654528d29d552352d3fddbff18)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>