qmltc: extend QmltcMethod with optional comments
The methods generated with Qmltc are facing the user:
therefore add the possibility to add internal comments
before generating the method call such that the user
can be warned or informed about certain unclear behaviors.

For example, the methods asked for in QTBUG-107570 needs to
warn the user about ownership issues on the arguments to
avoid dangling pointer references.

Task-number: QTBUG-107570
Change-Id: Id04ff7f9a36daad1da836bb421901b9b706a05e7
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>