Improve variable naming in QSGLayer implementation
'size' and 'rect' are names that should be avoided. In Qt 6 newer
code, such as QRhi, explicitly uses pixelSize instead of size, and
this is visible in some of the newer Qt Quick code as well, but old
code from Qt 5 times suffers from confusion still: even though on
the scenegraph level "everything" is in pixels, there are things
that aren't, for example the "rect" that gets passed to the
projection matrix calculation is in logical units! Comprehension
and maintenance gets challenged also when there are higher level
"size", "rect", "textureSize", etc. kind of names in QQuickItem
and elsewhere that are not necessarily in pixels yet can be
mistaken to map 1:1 to the scenegraph data.

Try to indicate this in argument and member names.

Also add a missing word to the QSGTexture docs. The overview
mentions this in fact but the function doc does not.

Task-number: QTBUG-107691
Change-Id: Idef7a83a2194ba5c0e3b12b1c0115b2b13fb7d66
Reviewed-by: Andy Nichols <andy.nichols@qt.io>