Restore the interpretation of user-provided layer/effect textureSize
ShaderEffectSource.textureSize and layer.size are meant to be in pixels.
For the latter this is documented. The former does not mention it since
it all comes from before high DPI times.

When the ShaderEffectSource gets an implicit size from an item, because
there is no user-provided textureSize (the most common case), that needs
to cater for high DPI of course, multiplying the item size with the
scale factor. But when having an explicitly set size, that is already
in pixels.

Fixes: QTBUG-107691
Change-Id: I375cac0e20b066288555302c91dc836d73bf9415
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>