Implement modal blocking for wheel events
The implementation of modal blocking for input events is distributed
over several Qt Quick classes. It's partially implemented by the overlay
handling mouse events for children (via childMouseEventFilter),
partially implemented via eventFilter, and partially happens to work if
the overlay is the top most item under the mouse.

These incomplete implementations only consider mouse and touch events.
QQuickDeliveryAgent redirects those explicitly to childMouseEventFilter.
It doesn't do anything for wheel or tablet events.

Since the overlay is the only data structure in Qt Quick that maintains
a list of popups (QQuickPopups add and remove themselves via addPopup),
the implementation of modality needs to live in QQuickOverlay. And since
the overlay installs itself as an event filter, it sees every input
event and can implement the modal blocking there.

For that it needs to know if the first item that would see the event is
modally blocked by any of the currently open popups. That is the case if
the popup is not an ancestor of the receiver item.

Add the case of nested modal popups to the existing QQuickPopup test.

Ultimately, the entire modality implementation should probably move into
this eventFilter, for all pointer events (key events are taken care of
by moving the focus when a popup opens and closes), in which case we
need to respect the auto-closing policy in QQuickPopup, and overlayEvent
overrides.

Fixes: QTBUG-107013
Pick-to: 6.4 6.2
Change-Id: I60359edad08586041b6f4080edd85be926cad091
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>