QQmlJSUtils: make a local array const, not constexpr
While QLatin1String is constexpr, and so one can create a constexpr
array of QLatin1String, there's really no point in doing so, because
if, indeed, the compiler creates a static lifetime object out of this,
it'll do nothing but create relocations, and one set per template
argument(!) at that.

By just making the array const, we allow the compiler to optimize this
into a bunch of lea and movi.

Also unbreaks the build with /Zc:lambda on MSVC.

Amends ca68b0cde7a3b9bad468dfae1935d5886f461499.

Change-Id: I9317c08df380cfe323efe5c09ffa7e688d6a6ffc
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>