Do not connect to a non-existent signal
Some "value" properties might not have a valueChanged signal, (e.g. the
HSV color picker in the non-native Color Dialog exposes a "value"
property, but "value" is really part of a compound property set, where
"color" is the main property. Therefore, it emits a "colorChanged()"
instead of "valueChanged()" signal).
Make the heuristic more fail-safe by checking for both a "value"
property and a "valueChanged()" signal.

From the meta-objects POV this heuristic is wrong (we could have asked
for QMetaProperty::hasNotifySignal() instead), but exposing *only* a
sub-component of a compound property doesn't make much sense normally,
so therefore this approach was chosen.

Also, a screen reader will reveal the real values of these
properties by drilling down to the respective input fields for H,S,V or
color.

Make the similar fix for the cursorPosition property.

Pick-to: 6.4
Task-number: QTBUG-106407
Change-Id: Ie7a4b0d62ac6011f9eccfe7405f0e3eb7a6cea10
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>