Keep track of glyphs we're asked to reference
If, when asked to populate the cache for a set of glyphs, you have only
(a) empty bounding rect(s), then we don't pre-render anything.
That's fine, but following 2fb43a35dfa08bb0746913e3363f1fcf21815e72
we deem such a cache as not active, since it has 0 glyphs in use
and 0 glyphs in the unused set. That's bad because we then decide to
invalidate the glyph cache to clean up resources, leading to a crash
later when we try to use it.

As a simple solution we keep track of glyphs we've been asked to
reference, and only count ourselves as inactive once we have no
glyphs left.

Fixes: QTBUG-107594
Change-Id: I7a276ac0b68eca2d9d49c55b29ef94d954baec89
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
(cherry picked from commit 3a7eb23c2adc41d456c55bd8d42444715122ed5f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>