ShaderEffect: Make source-destroyed connection more robust
As it stands now it would often end up crashing with inline-declared
sources. Pre-6.5 this used old-style connects with a different QObject
(the separate ShaderEffectImpl object no longer exists in dev) which
survived. The new approach with passing in the QQuickShaderEffect as
the connect's 'context' is apparently not safe due to the order
during destruction, leading to missing on destroyed() signals when
QObject deletes children (during the QQuickShaderEffect destruction
process) and thus ending up invalid objects in our source table. This
is also closer to what we did pre-6.5.

Fixes: QTBUG-107772
Change-Id: If1322b2525075af6e7091c6a58be8f2ad32a8587
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>