QQuickTableView: protect against delegate being deleted
While a delegate normally lives as long as the ItemView in which it is
used, that can unfortunately not be guarantueed. As demonstrated in the
test test case (and in a different way, in the bug report), it is
possible to delete the backing Component, and end up with a dangling
pointer in the table view.
Avoid this by using a QQmlGuard to reset the pointer to null when the
object gets deleted.

Pick-to: 6.4 6.2 5.15
Fixes: QTEXT-7
Change-Id: I6407b19a35b9707a039c9b4b3b387f191589cda8
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>