Ensure Flickable.flickEnded signal is emitted after flick-then-drag
hData and vData's flicking flags could get set to false in two places:
movementEnding(), which would also emit flickEnded(); or in
maybeBeginDrag(), which did not emit flickEnded(). So if the user did a
quick flick-flick-flick sequence, usually the flick would be restarted,
and movementEnding would emit the signal after motion stopped. But if
the user dragged too slowly the final time, the flick would not be
restarted, and movementEnding() did not emit flickEnded() because the
two flicking flags were already false. At the time the final drag
starts, we don't know if the user is going to drag fast enough to
qualify as a flick, so that's probably why maybeBeginDrag() doesn't
emit flickEnded(). But information is lost by setting the flicking
flags to false there.

So to make the flickStarted() and flickEnded() signals more symmetric,
we also need flags to remember that when the drag began, there was
residual momentum (it was already flicking). Then movementEnding() can
emit flickEnded() after movement has stopped, which is more consistent
with what the user should expect: all's well that ends well, even if it
started more than once.

It seems the intention was always that flickEnded() means the whole
sequence of flicks and drags had ended; so give a hint in the docs.

Declare the wasFlicking locals const, as a drive-by.

The bool bitfields in AxisData are converted to uint for proper packing
on MSVC.

Fixes: QTBUG-103620
Change-Id: I363aa3a0d95e1fff21b4b09f22e88e51b88c7c16
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit c9b8acba0e49a4d461183d069d12cbf18dfebe41)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>