Fix projection matrix to correct viewport size
When device pixel ratio is not 1.0, the projection matrix can
be incorrect because it uses floating point values when calculating
it, which can have fractional values.

Fix calculating projection matrix by rounding the logical pixel
size after dpr division. We don't render to fractional pixels so the
parameters to setProjectionMatrixToRect can't be fractional either.

Fixes white lines at the edges of the window.

Pick-to: 6.4
Fixes: QTBUG-95225
Change-Id: I85039ead5f31fb251c6cf760aa0b55b4b2a9e29d
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>