QQuickTableView: ensure that column sizes are in sync when using a sync view
When a TableView has a sync child, it will decide the row and column
sizes for it, so that the views are kept in sync. It can
happen, though, that the geometry of the sync view is smaller than
the geometry of the sync child. In that case, the sync child will
show more rows and column than the sync view.

This caused a bug when a TableView was placed inside a ScrollView
with a HorizontalHeaderView on top (as a sibling of the ScrollView).
In that case, the HorizontalHeaderView would actually be wider than
the TableView, since the latter would be shrinked a bit to make room
for a vertical scrollbar. The result would be that the
HorizontalHeaderView would end up showing more columns than TableView.
But Since TableView doesn't always know the width of a column that has
not yet been loaded, the size of the header column would also end up
wrong.

This patch will make sure that the "virtual" viewport of the sync
view is at least as large as the size of the sync children. This will
ensure that we end up filling up a TableView with the same amount of
rows and columns as any of the connected header views.

Fixes: QTBUG-107837
Change-Id: I32926738d48d6bd3b0892e6256a8f8d1ed5f1fa2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 32372f8895369bdd149fb90952b8054214a6dfe8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>