QQmlStandardPaths: don't convert URLs to local files
Some of the paths return by QStandardPaths are not
local files, but URLs with a different scheme. In
particular, on iOS the QStandardPaths::PicturesLocation
will return "assets-library://". So be more
careful before converting a standard path to
a file system path, and check up-front if the
path is already a valid URL.

(\sa https://doc.qt.io/qt-6/qfiledialog.html#setDirectory)

Task-number: QTBUG-107844
Change-Id: Ic57e88833d519f5df484215360ee7e8d23c0106a
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>