MenuBar: Implement Alt-key navigation
On Windows (and other systems where the platform theme says so), a press
and release of Alt moves focus to the menubar. This is implemented in
QMenuBar by handling the shortcut override for Alt to start listening
for events delivered throughout the QApplication, and then grabbing
focus on release unless another user interaction canceled the grab.

Implement equivalently for Qt Quick: if a shortcut override event
for Alt bubbles up to the window's contentItem, then wait for the
release and grab focus unless interrupted.

Fixes: QTBUG-107703
Change-Id: I73ee1d493d6f6f808fb39325d703a77fc5238e12
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>