Use fixed font, tighten tolerance in TextEdit largeTextObservesViewport
In tst_qquicktextedit::largeTextObservesViewport(), we needed tolerances
on the expected region to be rendered after scrolling by a particular
pixel delta, because there is so much variation in the default font
between platforms. Clearly, the solution is to use a specific font file;
so now there's a FontLoader to apply it to the TextEdit.
"tarzeau_ocr_a.ttf" is already present elsewhere in this git repository.

High-dpi scaling still seems to make some difference though (rounding
differences presumably), so we still need the tolerances, but they
don't need to be as broad as they did.

Amends 9db23e0e04906cf9ea33e23fa41f34955e5e6fe0

Task-number: QTBUG-103096
Change-Id: Ia45d8a93c0ef3d4bfe1fa3e00ecfcc3e89979815
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>