MenuBar: open menu when bar has focused and mnemonic key is pressed
When the MenuBar has focus and contains a &File item, then pressing F
should open the menubar. For keypresses with text and no modifier, scan
the list of menus to see if any of them has a mnemonic Alt+ pressed key.
If so, activate the item and open the menu.

Task-number: QTBUG-107703
Change-Id: Idab62aa709fce3318d85c946536a456d4fccbfd8
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>