Maintain relative stacking order of popups
When we show a popup, the popup's item is reparented to the overlay to
stack on top of the rest of the UI, and to allow the overlay to take
care of modal blocking.

By reparenting each popupItem to the overlay, we lose the relative
stacking order of popups that were previously not siblings. A parent
popup with a high z-value might end up covering popups items that belong
to popups that are actually its own children.

We cannot change the reparenting of popup items to the overlay without
breaking the existing modality implementation. However, we can make sure
that a popup that gets shown, and is a descendent of the popup currently
on top, is stacked above existing popups in the overlay, and has a
z-value at least as high as the current top-popup.

[ChangeLog][Qt Quick Controls][Popup] When showing a popup that doesn't
have an explicit z value set and is a child of the current top popup,
then the popup's item's z value will be set to the z-value of the
current top item.

Fixes: QTBUG-105741
Change-Id: I5d9e3c58699eca13cb2bbff9f783d4e9fd4a7c97
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>