QQuickTest::initView: set minimum window width and height if unset
When writing a new test, it's easy to forget to set the root item's
width and height, and that's ok on any desktop platform: the QPA plugin
or the window system will give it a reasonable minimum size anyway
(100x100 or so). But on some embedded platforms the minimum window
size is much smaller, such as 1x1; and then we can end up with CI
failures that take too long to troubleshoot. So if width or height
is not set (and thus 0), set it to 100.

Fixes: QTBUG-107865
Change-Id: If6c300d6ff43e3ba8d9142c550579ef7d58d87fb
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>